+++ components/organisms/TopNav/top-nav.tsx
@@ -89,9 +89,9 @@ const Nav = ({ className }: { className?: string }) => {
           <li className="hidden lg:inline">
             <Link
               className={`tracking-tight font-medium text-sm text-slate-700 hover:text-orange-500 flex gap-2 items-center transition-all ${getActiveStyle(
-                router.pathname === "/starsearch"
+                router.pathname === "/star-search"
               )}`}
-              href={"/starsearch"}
+              href={"/star-search"}
             >
               <img src="/assets/star-search-logo.svg" alt="Star Search logo" className="w-5 h-5" />
               <p className="mt-0.5">StarSearch</p>